requirements.txt
azure-functions
azure-data-tables

